Exa17 fixed and numerically compatible with Exa14. (#244)
* Exa17 fixed and numerically compatible with Exa14.

Added all missing fixes compared to Exa14, pdg bug and sorting by energy.

The compacting operation on the track container invalidated the queues of leaked track references. Now leaked tracks are extracted and copied by the transport kernels.

* Fix missing synchronization if no compaction happens.

* Allow for fractional numbers in the milliontrackslots command

* Removed debugging leftovers

* Added per event real time printout to Exa14 and Exa17

* Fixing hepmc reader for MT for Exa17 (ported from #208)

+++ examples/Example17/include/DetectorConstruction.hh
@@ -63,14 +63,14 @@ public:
   void SetBufferThreshold(int value) { fBufferThreshold = value; }
 
   // Set total number of track slots on GPU
-  void SetTrackSlots(int value) { fTrackSlotsGPU = value; }
+  void SetTrackSlots(double value) { fTrackSlotsGPU = value; }
 
   std::vector<G4String> &GetSensitiveGroups() { return fSensitive_group; }
 
 private:
   int fVerbosity{0};        ///< Actually verbosity for AdePT integration
   int fBufferThreshold{20}; ///< Buffer threshold for AdePT transport
-  int fTrackSlotsGPU{1};    ///< Total number of track slots allocated on GPU (millions)
+  double fTrackSlotsGPU{1}; ///< Total number of track slots allocated on GPU (millions)
   G4String fGDML_file;
   G4String fRegion_name;
   std::vector<G4String> fSensitive_volumes;
